fix: yScale issue
