+++ sigle/src/modules/analytics/stats/AreaChart.tsx
@@ -1,10 +1,12 @@
 import { AreaClosed, LinePath } from '@visx/shape';
 import { LinearGradient } from '@visx/gradient';
-import { curveNatural } from '@visx/curve';
-import { AxisScale } from '@visx/axis';
+import { curveBasis, curveNatural } from '@visx/curve';
+import { AxisBottom, AxisLeft, AxisScale } from '@visx/axis';
 import { theme } from '../../../stitches.config';
 import { WithParentSizeProps } from '@visx/responsive/lib/enhancers/withParentSizeModern';
 import { Group } from '@visx/group';
+import { StatsData } from '../../../types';
+import { format } from 'date-fns';
 
 const violet = theme.colors.violet5.toString();
 const green = theme.colors.green3.toString();
@@ -14,18 +16,14 @@ const greenStroke = theme.colors.green11.toString();
 
 // accessors
 const getDate = (d: StatsData) => new Date(d.date);
-const getValue = (d: StatsData) => d.value;
-
-interface StatsData {
-  value: number;
-  date: string;
-}
+const getViews = (d: StatsData) => d.value;
+const getVisits = (d: StatsData) => d.visits;
 
 interface AreaChartProps extends WithParentSizeProps {
   data: StatsData[];
-  color: 'purple' | 'green';
-  margin?: { top: number; right: number; bottom: number; left: number };
+  margin: { top: number; right: number; bottom: number; left: number };
   width: number;
+  height: number;
   top?: number;
   left?: number;
   yMax: number;
@@ -37,8 +35,8 @@ interface AreaChartProps extends WithParentSizeProps {
 export const AreaChart = ({
   margin,
   data,
-  color,
   width,
+  height,
   yMax,
   top,
   left,
@@ -48,39 +46,93 @@ export const AreaChart = ({
 }: AreaChartProps) => {
   if (width! < 10) return null;
 
-  const gradient = color === 'purple' ? violet : green;
+  const tickFormat = (d: any) => {
+    return format(d, 'dd/MM');
+  };
+
+  const fontFamily = '"Open Sans", sans-serif';
+  const axisColor = theme.colors.gray6.toString();
+  const axisLabelColor = theme.colors.gray11.toString();
+  const axisBottomTickLabelProps = {
+    textAnchor: 'middle' as const,
+    fontFamily,
+    fontSize: 12,
+    fill: axisLabelColor,
+  };
+  const axisLeftTickLabelProps = {
+    dx: '-0.25em',
+    dy: '0.25em',
+    fontFamily,
+    fontSize: 12,
+    textAnchor: 'end' as const,
+    fill: axisLabelColor,
+  };
 
   return (
-    <Group
-      style={{ border: '1px solid red' }}
-      left={left || margin?.left}
-      top={top || margin?.top}
-    >
+    <Group left={left || margin.left} top={top || margin.top}>
+      <LinearGradient
+        id={'purple-gradient'}
+        from={violet}
+        fromOpacity={0.7}
+        to={violet}
+        toOpacity={0}
+      />
       <LinePath
         data={data}
         x={(d) => xScale(getDate(d)) ?? 0}
-        y={(d) => yScale(getValue(d)) ?? 0}
-        stroke={color === 'purple' ? violetStroke : greenStroke}
+        y={(d) => yScale(getViews(d)) ?? 0}
+        stroke={violetStroke}
         strokeWidth={4}
         curve={curveNatural}
-      />
-      <LinearGradient
-        id={color === 'purple' ? 'purple-gradient' : 'green-gradient'}
-        from={gradient}
-        fromOpacity={0.5}
-        to={gradient}
-        toOpacity={0}
-        toOffset={0.75}
       />
       <AreaClosed<StatsData>
         data={data}
         x={(d) => xScale(getDate(d)) ?? 0}
-        y={(d) => yScale(getValue(d)) ?? 0}
+        y={(d) => yScale(getViews(d)) ?? 0}
         yScale={yScale}
-        fill={
-          color === 'purple' ? 'url(#purple-gradient)' : 'url(#green-gradient)'
-        }
+        fill={'url(#purple-gradient)'}
         curve={curveNatural}
+      />
+
+      <LinearGradient
+        id={'green-gradient'}
+        from={green}
+        fromOpacity={0.7}
+        to={green}
+        toOpacity={0}
+      />
+      <LinePath
+        data={data}
+        x={(d) => xScale(getDate(d)) ?? 0}
+        y={(d) => yScale(getVisits(d)) ?? 0}
+        stroke={greenStroke}
+        strokeWidth={4}
+        curve={curveBasis}
+      />
+      <AreaClosed<StatsData>
+        data={data}
+        x={(d) => xScale(getDate(d)) ?? 0}
+        y={(d) => yScale(getVisits(d)) ?? 0}
+        yScale={yScale}
+        fill={'url(#green-gradient)'}
+        curve={curveBasis}
+      />
+      <AxisBottom
+        top={yMax}
+        scale={xScale}
+        numTicks={7}
+        tickLabelProps={() => axisBottomTickLabelProps}
+        stroke={axisColor}
+        tickStroke={axisColor}
+        tickFormat={tickFormat}
+        hideTicks={true}
+      />
+      <AxisLeft
+        scale={yScale}
+        numTicks={5}
+        stroke={axisColor}
+        tickStroke={axisColor}
+        tickLabelProps={() => axisLeftTickLabelProps}
       />
       {children}
     </Group>
